Change migration and factory

+++ database/migrations/2024_01_06_110303_create_products_table.php
@@ -13,13 +13,13 @@ return new class extends Migration {
     {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
-            $table->string("product_name");
-            $table->longText("description");
-            $table->double("price", 11, 2, true)->nullable();
+            $table->string('product_name');
+            $table->longText('description');
+            $table->decimal('price', 11, 2, true)->nullable();
             $table->bigInteger('stock_quantity')->default(0);
-            $table->bigInteger("category_id")->nullable();
-            $table->string("slug")->nullable();
-            $table->tinyText("status")->default(ProductStatus::ACTIVE->name);
+            $table->bigInteger('category_id')->nullable();
+            $table->string('slug')->nullable();
+            $table->tinyText('status')->default(ProductStatus::ACTIVE->name);
             $table->json('json')->nullable();
             $table->timestamps();
             $table->softDeletes();
